requirements.txt
#dependencies

torch==2.0.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
tensorflow-cpu>=2.13.0
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=21.0.0
